Use psr-4 directory structure
