+++ src/Controllers/ImportExport.php
@@ -2,9 +2,9 @@
 
 namespace IgniterLabs\ImportExport\Controllers;
 
-use Admin\Classes\AdminController;
-use Admin\Facades\AdminMenu;
-use Admin\Facades\Template;
+use Igniter\Admin\Classes\AdminController;
+use Igniter\Admin\Facades\AdminMenu;
+use Igniter\Admin\Facades\Template;
 use Igniter\Flame\Exception\ApplicationException;
 use IgniterLabs\ImportExport\Classes\ImportExportManager;
